feat: add modal to watch the video button
